+++ libs/index.js
@@ -13,6 +13,8 @@ import updateNotNewsByMenuId from './updateNotNewsByMenuId';
 import updatePersonalize from './updatePersonalize';
 import updateHotKeywords from './updateHotKeywords';
 import updateInstantNews from './updateInstantNews';
+import updateNewestNews from './updateNewestNews';
+import updateNewestImages from './updateNewestImages';
 import refreshIndexPage from './refreshIndexPage';
 import parseRssFeed from './parseRssFeed';
 
@@ -31,6 +33,8 @@ module.exports = {
     updatePersonalize,
     updateHotKeywords,
     updateInstantNews,
+    updateNewestNews,
+    updateNewestImages,
     refreshIndexPage,
     parseRssFeed
 };