+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 NOWnews-api change logs
 
+## 1.0.3 - 2017-07-04
+### Added
+-  增加 `/news/newest` 端點 @SimonSun
+-  增加 `/news/images` 端點 @SimonSun
+
+### Changed
+- Image model 的 virtual 欄位，加入判斷是否為 XXX.nownews.com 的圖片來源
+
 ## 1.0.2 - 2017-07-03
-### Change
+### Changed
 -  更新get /news資料庫端點，為每筆news加上pageviews。(api-admin) @appleoxxo
 
 ## 1.0.1 - 2017-07-03
-### Change
+### Changed
 - 更新 china 資料庫端點 @esbb48
 
 ### Fixed